fix compiler errors
Former-commit-id: f8b8ad33042c678e8d2409732de6571f52a7315d [formerly 5d2b632529895a2a6e0e43ce6f2b0e19a4c6139a]
Former-commit-id: 1adfa2c51c8007ff6694da7df0fe9d92220a099a

+++ accelerator/cpu/image/image_mixer.cpp
@@ -161,7 +161,7 @@ public:
 
 		convert(items, format_desc.width, format_desc.height);
 
-		auto result = spl::make_shared<buffer>(format_desc.size, 0);
+		auto result = spl::make_shared<buffer>(format_desc.size, static_cast<uint8_t>(0));
 		if(format_desc.field_mode != core::field_mode::progressive)
 		{
 			draw(items, result->data(), format_desc.width, format_desc.height, core::field_mode::upper);
